Function trezor_client::unique

pub fn unique(debug: bool) -> Result<Trezor>

Try to get a single device. Optionally specify whether debug should be enabled or not. Can error if there are multiple or no devices available. For more fine-grained device selection, use find_devices(). When using USB mode, the device will show up both with debug and without debug, so it’s necessary to specify the debug option in order to find a unique one.