Trait trezor_client::TrezorMessage

pub trait TrezorMessage: Message + Default {
    const MESSAGE_TYPE: MessageType;
}

This trait extends the protobuf Message trait to also have a constant for the message type code. This getter is implemented in this file for all the messages we use.

Associated Constants

const MESSAGE_TYPE: MessageType